refactore: modify the thme switcher folder
