+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react'
 import './App.css'
-import ThemeSwitcher from './components/ThemeSwitcher.tsx'
+import ThemeSwitcher from './components/ThemeSwitcher/ThemeSwitcher.tsx'
 
 function App() {
   const [themeApp, setThemeApp] = useState("light");
